cherry-picks
Implement oculus-go-controls

This is mostly a copy and rename of the GearVR controls. This is to get everything boot strapped for now and then diverge them later while consolidating common functionality into a set of helpers.

The same for the tests, I've done a bit more consolidation in them, but also this is a copy.

For the controller we rely on the existing Gear VR controllers as well until we are able to upload Oculus Go specific controllers.

Add Oculus Go controller model with button highlighting support

Address CR changes:

1. This assumes a new version of the GLTF model that has separate materials for the trackpad and trigger, so the code to clone the material has been removed.
2. Tests folder fixed to tests
3. Separate buttons now referenced by name instead of array index.

add util function to detect Oculus Go user-agent

fix oculus go lint